Update mutation to use GetOrCreateUser
GetOrCreateUser takes a username and address and returns a user
