+++ packages/server/src/graphql/schema/index.ts
@@ -1,8 +1,8 @@
 import { GraphQLSchema } from "graphql";
 
-import MutationType from "./type/MutationType";
-import QueryType from "./type/QueryType";
-import SubscriptionType from "./type/SubscriptionType";
+import { MutationType } from "./type/MutationType";
+import { QueryType } from "./type/QueryType";
+import { SubscriptionType } from "./type/SubscriptionType";
 
 export const schema = new GraphQLSchema({
   query: QueryType,
