+++ packages/server/src/graphql/schema/type/SubscriptionType.ts
@@ -3,7 +3,7 @@ import { GraphQLObjectType } from "graphql";
 // import ChallengeSubscriptions from '../../poll/subscriptions';
 import CommentSubscriptions from "../../modules/comment/subscriptions";
 
-export default new GraphQLObjectType({
+export const SubscriptionType = new GraphQLObjectType({
   name: "Subscription",
   fields: {
     // ...ChallengeSubscriptions,
